Add endpoint to update accounts

+++ src/main/java/ch/raising/services/AccountService.java
@@ -3,9 +3,12 @@ package ch.raising.services;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -14,6 +17,7 @@ import org.springframework.stereotype.Service;
 
 import ch.raising.models.Account;
 import ch.raising.models.AccountDetails;
+import ch.raising.models.AccountUpdateRequest;
 import ch.raising.models.Response;
 import ch.raising.data.AccountRepository;
 
@@ -65,9 +69,15 @@ public class AccountService implements UserDetailsService {
      * @param id the id of the account to delete
      * @return ResponseEntity with status code and message
      */
-    public ResponseEntity deleteAccount(int id) {
-        accountRepository.delete(id);
-		return ResponseEntity.ok(new Response("Successfully deleted account"));
+    public ResponseEntity<?> deleteAccount(int id) {
+        if(accountRepository.find(id) == null)
+            return ResponseEntity.status(500).body(new Response("Account doesn't exist"));
+        try {   
+            accountRepository.delete(id);
+            return ResponseEntity.ok(new Response("Successfully deleted account"));
+        } catch (Exception e) {
+            return ResponseEntity.status(500).body(new Response(e.getMessage()));
+        }
     }
 
     /**
@@ -87,5 +97,45 @@ public class AccountService implements UserDetailsService {
      */
     public Account findById(int id) {
         return accountRepository.find(id);
+    }
+
+    /**
+     * Check if request comes from own account (or admin account)
+     * @param id the id of the account to check against
+     * @param request instance of the http request
+     * @return true if account belongs to request, false otherwise
+     */
+    public boolean isOwnAccount(int id, HttpServletRequest request) {
+        Account account = findById(id);
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        boolean isAdmin = request.isUserInRole("ROLE_ADMIN");
+        if(account == null)
+            return false;
+
+        if(!account.getUsername().equals(username) && !isAdmin)
+            return false;
+
+        return true;
+    }
+
+    /**
+     * Update user account
+     * @param id the id of the account to be updated
+     * @param 
+     * @return Response entity with status code and message
+     */
+    public ResponseEntity<?> updateAccount(int id, AccountUpdateRequest req) {
+        if(accountRepository.find(id) == null)
+            return ResponseEntity.status(500).body(new Response("Account doesn't exist"));
+        try {  
+            if(req.getUsername() != null) {
+                if(accountRepository.findByUsername(req.getUsername()) != null)
+                    return ResponseEntity.status(500).body(new Response("Username already in use"));
+            }
+            accountRepository.update(id, req);
+            return ResponseEntity.ok(new Response("Successfully updated account"));
+        } catch (Exception e) {
+            return ResponseEntity.status(500).body(new Response(e.getMessage()));
+        }
     }
 }
